adds more test for added custom hooks
